fix the four CI failures this wave introduced or left standing
An empty bar panel raised where the plotly twin rendered: the tick budget indexed positions with
np.asarray(_keep), and an empty keep-list is a float64 array, which numpy refuses as an index. A
zero-category panel is a real input -- the degenerate-input matrix renders one -- so the index dtype is
explicit now.

plotly.py had grown to 1137 LOC, over the 1000-LOC house limit. _line and the marker-label helpers it needs
are carved into _plotly_line.py and bound back onto the class, the same pattern the heatmap and network
families already use; the parent is 905. The carve immediately hit the failure mode that rule exists for --
seven names that resolved only in the parent -- caught by ruff F821 and fixed with the same call-time
import the heatmap sibling uses for the cycle.

test_fs_hybrid_manifest parsed a manifest with stdlib json purely to check key ordering; orjson does that,
so it needs no whitelist entry.

The base-leakage guard tests passed discovery=None, but the guard records _base_leakage_guard_ran_ on that
object BEFORE the early return, precisely so "ran and found nothing" stays distinguishable from "never
ran", and the single production caller passes self. They use a stand-in object now, and assert the flag in
both directions -- which the docstring implies and nothing was checking.

+++ src/mlframe/reporting/renderers/matplotlib.py
@@ -775,7 +775,7 @@ class MatplotlibRenderer:
             _cats = [truncate_bar_label(c, keep_tail=p.label_keep_tail) for c in p.categories]
             _keep = _thin_tick_positions(n_cat, ticks_that_fit(_measured_axis_in(ax, horizontal=True), n_cat,
                                                                pitch_in=rotated_tick_pitch_in(_HEATMAP_TICK_FONTSIZE, 0)))
-            ax.set_yticks(pos[np.asarray(_keep)])
+            ax.set_yticks(pos[np.asarray(_keep, dtype=np.int64)])
             ax.set_yticklabels([_cats[i] for i in _keep], fontsize=_HEATMAP_TICK_FONTSIZE)
             ax.invert_yaxis()  # first category on top -> worst-first ranking reads top-down
             _pad_sparse_category_axis(ax, len(p.categories), horizontal=True)
@@ -804,7 +804,7 @@ class MatplotlibRenderer:
                     _rotation = 45.0
             _pitch = rotated_tick_pitch_in(_HEATMAP_TICK_FONTSIZE, _rotation) if _rotation else label_width_pitch_in(_cats_v, _HEATMAP_TICK_FONTSIZE)
             _keep_v = _thin_tick_positions(n_cat, ticks_that_fit(_axis_in, n_cat, pitch_in=_pitch))
-            ax.set_xticks(pos[np.asarray(_keep_v)])
+            ax.set_xticks(pos[np.asarray(_keep_v, dtype=np.int64)])
             ax.set_xticklabels(
                 [_cats_v[i] for i in _keep_v],
                 rotation=_rotation,
